+++ src/TrelloDotNet/TrelloClient.Memberships.cs
@@ -1,8 +1,11 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using TrelloDotNet.Control;
 using TrelloDotNet.Model;
+using TrelloDotNet.Model.Options;
+using TrelloDotNet.Model.Options.GetBoardOptions;
 
 namespace TrelloDotNet
 {
@@ -30,6 +33,60 @@ namespace TrelloDotNet
         {
             await _apiRequestController.Put($"{UrlPaths.Boards}/{boardId}/memberships/{membershipId}", cancellationToken, 0,
                 new QueryParameter("type", membershipType.GetJsonPropertyName()));
+        }
+
+        /// <summary>
+        /// Get the Workspace and Board Memberships for current Token User (It is assumed that if user is Admin on Workspace they are also Admin on Boards underneath that workspace)
+        /// </summary>
+        /// <returns></returns>
+        public async Task<TokenMembershipOverview> GetCurrentTokenMembershipsAsync()
+        {
+            Member member = await GetTokenMemberAsync();
+            var organizations = await GetOrganizationsCurrentTokenCanAccessAsync();
+            var boards = await GetBoardsCurrentTokenCanAccessAsync(new GetBoardOptions
+            {
+                BoardFields = new BoardFields(BoardFieldsType.Name),
+                Filter = GetBoardOptionsFilter.Open
+            });
+            var organizationMemberships = new Dictionary<Organization, MembershipType>();
+            var boardMemberships = new Dictionary<Board, MembershipType>();
+
+            foreach (Organization organization in organizations)
+            {
+                Membership orgMemberShip = organization.Memberships.FirstOrDefault(x => x.MemberId == member.Id);
+                if (orgMemberShip != null)
+                {
+                    organizationMemberships.Add(organization, orgMemberShip.MemberType);
+                    if (orgMemberShip.MemberType == MembershipType.Admin)
+                    {
+                        //Since user is workspace admin, they are automatically also board admin for all boards under
+                        foreach (var boardId in organization.BoardIds)
+                        {
+                            Board board = boards.FirstOrDefault(x => x.Id == boardId);
+                            if (board != null)
+                            {
+                                boardMemberships.Add(board, MembershipType.Admin);
+                            }
+                        }
+                    }
+                }
+            }
+
+            foreach (Board board in boards.Where(board => !boardMemberships.ContainsKey(board)))
+            {
+                var membershipsOfBoard = await GetMembershipsOfBoardAsync(board.Id);
+                Membership boardMemberShip = membershipsOfBoard.FirstOrDefault(x => x.MemberId == member.Id);
+                if (boardMemberShip != null)
+                {
+                    boardMemberships.Add(board, boardMemberShip.MemberType);
+                }
+            }
+
+            return new TokenMembershipOverview
+            {
+                OrganizationMemberships = organizationMemberships,
+                BoardMemberships = boardMemberships
+            };
         }
     }
 }